Modularize pool kernels. (#4267)
DEV
